Update visualize_splitted_documents code
in addition to the arguments that visualize_documents takes, this function also takes a
splitted_docs list to indicate which docs originate from the same document and
the bool hide_unassigned_snippets_from_assigned_docs to indicate, if docs not assigned to a topic,
that originates from a document, where other snippets have been assigned to topics, should be
displayed.

The docs assigned to the same topic, which originates from the same document, will have their
coordinates averaged.